Move auth types to the CLI and remove dead code
